move check_max_queries_dict to utils
